+++ src/lib/types/Client.d.ts
@@ -3,12 +3,21 @@ import { Command } from "../exec/Command";
 import { AutoResponse } from "../types/AutoCommand";
 import UserManager from "../manager/UserManager";
 import ServerManager from "../manager/ServerManager";
+import { ScheduledScript } from "./ScheduledScripts";
 
 export interface Client extends Client {
     isReady: boolean,
     commands: Collection<string, Command>,
     autoResponses: Collection<string, AutoResponse>,
 
+    scheduled: Array<ScheduledScript>,
+
     userManager: UserManager,
-    serverManager: ServerManager
+    serverManager: ServerManager,
+
+    cache: {
+        prefix: {
+            [guild: string]
+        }
+    }
 }