+++ src/lib/types/ScheduledScripts.ts
@@ -1,0 +1,6 @@
+import { Client } from "./Client";
+
+export interface ScheduledScript {
+    every: number;
+    run: (client: Client) => Promise<void>;
+}