+++ src/preload/commands/commands/index.ts
@@ -16,9 +16,9 @@ export default class Commands extends Command {
         let embed = new RichEmbed();
         embed.setAuthor("TFBot", client.user.avatarURL);
         embed.setTitle(`List of all commands`);
-        embed.addField("Commands",(client.commands.filter(cmds => !cmds.requiresRoot && !cmds.name.includes("api")).map(cmd => cmd.name).join(', ')));
+        embed.addField("Commands",(client.commands.filter(cmds => !cmds.requiresRoot && !cmds.name.includes("pugscrim")).map(cmd => cmd.name).join(', ')));
         embed.addField("Automatic Commands", client.autoResponses.map(cmd => cmd.name).join(', '));
-        embed.setFooter(`\nYou can say ${config.PREFIX}help [command name] to get info on a specific command!`);
+        embed.setFooter(`\nYou can say ${await this.getPrefix(msg.guild.id)}help [command name] to get info on a specific command!`);
         embed.setColor(16098851);
         msg.channel.send(embed);
         return true;